refactor impl Display for Table
This passes refs to table and path around, and avoids the use
of Cell and RefCell. It splits the job of walking the tree from
that of printing bytes by using the visitor pattern.

This is probably net-neutral for readability, but it allows
us to extend the formatting logic later if we need to.