refactor!: Reduce Repr::new_unchecked escape hatches
To help ensure we don't mix up what we are deaing with, this involves an
audit of `Repr::new_unchecked` and tries to move the calls to it closer
to where validation happens.

As part of this, this removes the "magic" of table keys being either
Rust strings or toml keys.

BREAKING CHANGE: Can no longer use TOML keys when mucking with tables.