fix: Ensure floats stay as floats
Relying on Rust to format our floats was giving us "1000" which we then
read as an integer, failing tests.

This gives us a cheap-to-implement solution so we'll always preserve the
float-ness, even if it can be ugly.  We can iterate if this is a
problem.  I just don't expect many floats so I don't expect this to be a
problem.

While at it, this changed the nan issue (#169) to be fixed at the
source, at repr generation.