perf: Skip table lookup per entry
Before, for every key-value pair, we'd find the table and insert.  Now,
we store off the table as we come across each header and insert directly
into it (ignoring the lookups for dotted keys).

I was hoping this would drop us from N lookups to 1 lookup per table but
we need 2 lookups per table to ensure proper error reporting.