fix: Sane default array decor (#207)
When porting cargo-edit to toml_edit 0.3, I found the default array
formatting is off.  This is because our `Display` doesn't have
knowledge of its context.  Generally, the caller knows better what the
default decode should be.

This adds a new trait for encoding which allows the caller to pass in
the needed context.