perf: small-string optimization
For a `key = "value"`, we at least allocate
- `"key"` for the repr
- `"key"` for being able to reuse logic around repr
- `"key"` for the hash table
- ` ` before `=`
- ` ` after `=`
- `"value"` repr
- `"value"` value

Compare this to what `toml-rs` has to deal with
- `"key"`
- `"value"`

This drops the `medium` cargo benchmark from 250us to 203us.  Still not
as big of a drop as I'm hoping for but there are still a lot of spurious
allocations or slow operations we do, like
- Allocations when calling into `Index`
- Allocations in various table functions

I considered making this optional but its such a small dependency, I
figured, eh.

I chose `kstring` over
- `smartstring` because we rarely need mutability and the assumptions it
  makes scares me from using it in anything but a top-level `[[bin]]`
- `compact_str` didn't seem to have benchmarks last I looked
- `smol_str` last benchmarks I ran, it looked to be slower iirc