refactor!: Make InternalString a newtype (#210)
This both helps clarify where and why we use this and allows us to
change the implementation, like using smalll-string optimization.

As part of this, I made it immutable.

For now, I left this more "internal" and it only shows up in APIs for
`Into<InternalString>` with the one exception being `IntoIter` for
tables.  Depending on performance and API friendliness, we can consider
changing `Value`s string type.

BREAKING CHANGE: String type changed in a few places.