perf: Switch to byte processing
The grammar is fairly neutral, so long as we can handle "non-ASCII".

This does introduce `unsafe` for cases where the grammar guarantees
we'll have valid UTF-8, to avoid validation.  There might be some more
cases where we can bypass the checks; I was being cautious and only
using `unsafe` for 7-bit ASCII.  If we ensure our input is only UTF-8
and we only split on 7-bit ASCII, we are probably safe to use bypass
validation everywhere.

So this saves us having to encode UTF0-8 as `char` at the cost of some
UTF-8 validation.  Looks like this gives us a 6-12% reduction in parsing
time.