fix!: Remove serde ambiguity with Datetime
We've been serializing datetimes as strings, making them it so on every
string, we have to try both types.  By following `toml-rs`s approach, we
remove this ambiguity and make it easier to deserialize.

BREAKING CHANEG: We no longer serialize/deserialize datetimes as strings
but instead as an internal type.  To use dates, you must use
`toml_edit::Datetime`.