feat(serde): Support `to_item` (#246)
Easily convert any serializable type into `toml_edit::Item`.  This is
needed by cargo to serialize some data into the lock file.  cargo might
also need `to_string` to handle any type, but I'm trying to hold off on
that as I prefer `to_string` to be "strictly typed".  By that I mean
that the output is always considered a TOML document, rather than being
either a TOML document or a fragment of a TOML document.

As part of this, we now have
- `InlineTable::into_table`
- `Item::into_table` now also converts `InlineTable` into a table.
- `table.into::<Document>()`