feat: `*Table::sort_values_by` to support custom sorting (#276)
We have large TOML documents that have been sorted using a
non-TOML-aware sorter (for example, highlighting a block of code and
using VSCode to sort a table). This ends up sorting quoted keys ahead of
non-quoted keys.

    [table]
    "a key" = 1
    "b key" = "Example"
    "c key" = false
    akey = 2
    bkey = "Another example"
    ckey = true

When adding a new entry to a table, we would like to keep that sort. The
current `Table::sort_values` function intelligently sorts by the value
of the key.

    [table]
    "a key" = 1
    akey = 2
    "b key" = "Example"
    bkey = "Another example"
    "c key" = false
    ckey = true

To allow controlling the sort order, add `Table::sort_values_by` (which is a
thin wrapper over `IndexMap::sort_by`). This allows us to do the
following to preserve non-intelligent sorting.

    table.insert("another", Item::Value("Another".into()));
    table.sort_by(|k1, _, k2, _| k1.to_repr().as_raw().cmp(k2.to_repr().as_raw());

And get the output that we want.

    [table]
    "a key" = 1
    "b key" = "Example"
    "c key" = false
    akey = 2
    another = "Another"
    bkey = "Another example"
    ckey = true