fix(ser): Render empty tables when using `to_string_pretty`
I considered making it so we always showed empty implicit tables but I
figured that in the editing case, we could show tables that were
originally hidden just by deleting entries.  This seems surprising for a
user and they shouldn't have to account for it.

So I limited the fix to `to_string_pretty`.

This was found by rust-lang/cargo#10349.