fix(edit): No newline when first table is dotted
When making a table pretty, we try to put newlines between tables.
Since the decor is tracked in each table, we have to protect against
that newline being at the start of the document.  Before, we only
checked if the first item in the document was a table but we didn't
check if that table is visible.  Now we watch for the first visible
table for removing the newline.