fix(de): Track the keys as an alternative to spans
When adding spans, the loss of key tracking was fine as that was covered by the
span.  In testing this in cargo, I found there were too many cases where spans
were removed (e.g. config subsystem converting to `toml::Value` and then
converting to desired type), so adding it back in.