fix(parser): Ensure error spans are valid for escape sequences
After a `\` in a string, the parser consumes the next character and
then validates it.
If the validation fails, it means that the parser is now pointing to the
next byte after it which could be within a byte sequence.

We likely want to fix up some of these cases so errors point to the
right thing but in the mean time, we should make sure this doesn't
produce a bad span.