feat: Support formatting hex/oct/bin integers
Add methods that create `Formatted<i64>` in hexadecimal, octal, and
binary representation. In accordance with TOML 1.0 spec, these forms can
be created only if the integer is nonnegative. For hexadecimal, it is
possible to create either uppercase or lowercase representation.

This commit does not change the AST and should be fully backwards-
compatible.