feat(toml): Add DeTable::parse_recoverable
I had considered naming this new method `parse` and the `Result` method
`try_parse` but the `Result` method should be the default / go-to for
people and so the shorter name that matches similar functions is a
better choice.

`recovery` felt odd.

Having it after `parse`, instead of `recoverable_parse`, keeps the
emphasis on the most important part.

Fixes #371